requirements.txt
django==2.0.1
dj-database-url==0.5.0
django-storage-url==0.5.0
whitenoise==5.2.0
boto3==1.14.49
psycopg2==2.8.5
uwsgi==2.0.19.1

django-cms==3.7
djangocms-text-ckeditor
django-filer
djangocms-bootstrap4
djangocms-snippet==3.0.0
djangocms-style
djangocms-file==3.0.0
